move translation tests to renderer
